Work around status management
Lots of work to better understand when the mixer is available or not
Mixer should not update availability properly
Data updates should be slightly better, caterng for some of the edge
cases.

Added in new sensor for firmware version.
Potentially breaking change where the name of the entity can now be set
properly on setup, it defaults to the network name of the mixer but then
gives the user the chance to change it.  This better to use than the
random UUID that was initially setup.